스프링부트3 실습코드 3

+++ FirstObject/src/main/java/com/example/firstobject/entity/Member.java
@@ -21,6 +21,6 @@ public class Member {
     @Column
     private String email;
     @Column
-    private int password;
+    private Long password;
 
 }
